move everything to a workspace
