+++ Cargo.toml
@@ -1,25 +1,7 @@
-[package]
-name = "gerritbot"
-version = "0.7.0"
-authors = ["boxdot <d@zerovolt.org>"]
-license = "MIT/Apache-2.0"
-edition = "2018"
-
-[dependencies]
-docopt = "1.0"
-futures = "0.1"
-gerritbot-gerrit = { path = "gerritbot-gerrit" }
-gerritbot-spark = { path = "gerritbot-spark" }
-itertools = "0.8"
-lazy_static = "1.3"
-log = "0.4"
-lru_time_cache = "0.9"
-regex = "1.1"
-rlua = "0.16"
-rusoto_core = "0.36"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-serde_yaml = "0.8"
-shellexpand = "0.1"
-stderrlog = "0.4"
-tokio = "0.1"
+[workspace]
+members = [
+  "gerritbot-bot",
+  "gerritbot-gerrit",
+  "gerritbot-spark",
+]
+default-members = ["gerritbot-bot"]
