+++ gerritbot-bot/src/bin/gerritbot.rs
@@ -9,9 +9,8 @@ use log::{debug, error, info, warn};
 
 use gerritbot_gerrit as gerrit;
 use gerritbot_spark as spark;
-
-mod args;
-mod bot;
+use gerritbot as bot;
+use gerritbot::args as args;
 
 /// Create spark message stream. Returns a future representing a webhook server
 /// and a stream of messages.
